LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/uievents/mouse/mouseenter-mouseleave-on-drag-expected.txt


FAIL Test dragging across inner div assert_equals: Move to origin: Event  mismatch expected "mouseout@outer" but got "mousemove@outer"
FAIL Test dragging into inner div assert_equals: Move over outer element: Event  mismatch expected "mouseover@outer" but got "mouseout@outer"
FAIL Test dragging out of inner div assert_equals: Move over inner element: Event  mismatch expected "mouseover@inner" but got "mouseleave@outer"
